requirements.txt
# Python dependencies for Wildlife Camera Trap Detection System

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
streamlit>=1.28.0

# Dataset Download
kaggle>=1.5.16

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Image Augmentation
albumentations>=1.3.0

# Metrics and Evaluation
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
